apps/nba-season-backfill/pyproject.toml
[tool.poetry]
name = "nba-season-backfill"
version = "0.1.0"
description = "Containerized NBA 2024-25 season data backfill application"
authors = ["Hoopstat Haus Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
hoopstat-observability = {path = "../../libs/hoopstat-observability", develop = true}
nba-api = "^1.5.1"
boto3 = "^1.35.96"
pyarrow = "^19.0.0"
click = "^8.1.7"
tenacity = "^9.0.0"
pydantic = "^2.10.8"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
ruff = "^0.12.4"
black = "^25.1.0"
pytest-cov = "^6.2.1"
pytest-mock = "^3.14.0"

[tool.poetry.scripts]
backfill = "app.main:main"
test = "app.scripts:test"
lint = "app.scripts:lint"
format = "app.scripts:format_code"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Ruff configuration
[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.lint.mccabe]
max-complexity = 10

# Black configuration
[tool.black]
target-version = ["py312"]
line-length = 88

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "--strict-markers --disable-warnings"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]